Apply stashed changes

+++ database/migrations/2025_03_09_062732_create_model_has_permissions_table.php
@@ -13,14 +13,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('model_has_permissions', function (Blueprint $table) {
-            $table->unsignedBigInteger('permission_id');
-            $table->string('model_type');
-            $table->unsignedBigInteger('model_id');
-            $table->primary(['permission_id', 'model_id', 'model_type']);
-            $table->index(['model_id', 'model_type']);
-            $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('model_has_permissions')) { 
+            Schema::create('model_has_permissions', function (Blueprint $table) {
+                $table->unsignedBigInteger('permission_id');
+                $table->string('model_type');
+                $table->unsignedBigInteger('model_id');
+                $table->primary(['permission_id', 'model_id', 'model_type']);
+                $table->index(['model_id', 'model_type']);
+                $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -30,6 +32,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('model_has_permissions');
+        if (Schema::hasTable('model_has_permissions')) { // Kiểm tra trước khi xóa
+            Schema::dropIfExists('model_has_permissions');
+        }
     }
 };
